Split noddity-template-transformer.js into two files
